Struct rotor_tools::timer::Ticker

pub struct Ticker<M: Timer> {
    // some fields omitted
}

Ticker state machine

The structure implements rotor::Machine but exposes a simpler protocol that has just one method which is called when timer expires.

The Ticker machine also ensures that there are no spurious events.

Methods

impl<T: Timer> Ticker<T>

fn new(scope: &mut Scope<T::Context>, machine: T) -> Response<Ticker<T>, Void>